requirements.txt
clickhouse-connect==0.7.0
python-dotenv==1.0.0
google-generativeai==0.3.2
streamlit==1.28.1
pandas==2.1.3
urllib3<2.0.0
google-cloud-speech==2.21.0
SpeechRecognition==3.10.0
pyaudio==0.2.11
tabulate==0.9.0
